Backend/pyproject.toml
[project]
name = "sifter-kys-backend"
version = "0.1.0"
description = "Solana token analytics platform — caller analysis, wallet tracking, Elite 100 leaderboards"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "flask>=3.0.0,<4.0.0",
    "flask-compress>=1.17,<2.0.0",
    "flask-cors>=4.0.0,<5.0.0",
    "flask-limiter>=3.8.0,<4.0.0",
    "gunicorn>=21.2.0,<22.0.0",
    "gevent>=23.9.0,<25.0.0",
    # HTTP / async
    "requests>=2.31.0,<3.0.0",
    "aiohttp>=3.9.0,<4.0.0",
    # Auth & config
    "python-dotenv>=1.0.0,<2.0.0",
    "PyJWT>=2.10.1,<3.0.0",
    # Supabase
    "supabase>=2.27.0",
    # Redis & Celery
    "redis>=5.0.0,<6.0.0",
    "celery[redis]>=5.3.0,<6.0.0",
    "apscheduler>=3.10.0,<4.0.0",
    # Twitter
    "tweepy>=4.14.0,<5.0.0",
    # Analytics
    "duckdb>=1.0.0",
    "clickhouse-connect>=0.7.0",
    "networkx>=3.2.0,<4.0.0",
    "numpy>=1.26.0,<3.0.0",
    "python-louvain>=0.16,<1.0.0",
    "pandas>=2.1.0,<3.0.0",
    # CLI output
    "rich>=13.0.0,<14.0.0",
    "markdown-it-py>=3.0.0,<4.0.0",
    "structlog>=25.5.0",
    # OpenTelemetry (opt-in via OTEL_ENABLED=true)
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-flask>=0.41b0",
    "opentelemetry-instrumentation-requests>=0.41b0",
    "opentelemetry-instrumentation-redis>=0.41b0",
    "opentelemetry-instrumentation-celery>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "flake8>=6.0.0",
    "pyright>=1.1.0",
    "black>=24.0.0",
    "pip-audit>=2.7.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.uv]
# Application mode — no build system needed, just dependency management
package = false
